Changes made in tables in sqlite db
- in User class, renamed book_requests to books_requested and books_loans to books_read.
- In Section class, just deleted the function to calculate avg rating of the section. In Book class, added attribute is_issued which is a Boolean valued column with default = False

- **Main change - deleted (commented out) BookLoan class and now BookRequest class has everything.**

- In BookRequest class the original core attributes id, bookid and userid are unchanged. also added attributes with string datatype are unchanged: book_title and username.

For date requested sticking with just date and no timestamp. And the date_requested column is for storing date only no time stamp and default value is the current date when the request is made.
Important: renamed boolean attribute fulfilled to issued if librarian approves request. Default = False.

These two relationship related attributes are subject to change but for now attributes are of datatype string with 10 characters. date_issued and date_due. date issued is the current date when the book request is approved and the book request approve controller will fetch the current date using datetime module. Then data_due is calculated by adding 7 days to the date_issued using timedelta module.

- In BookRequest class, added a new attribute is_returned which is a Boolean valued column with default = False.

- Renamed BookFeedback class to BookRatting class. Also added the usual columns: id userid bookid book_title username attributes. Also added a new attribute rating which is an integer valued column with default = 0. this is the rating given by the user to the book. The default value is 0 and the user can give a rating between 1 to 5.